pyproject.toml
[tool.poetry]
name = "bbbs"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^3.2"
djangorestframework = "^3.12"
djangorestframework-camel-case = "^1.2"
drf-yasg = "^1.20"
djangorestframework-simplejwt = "^4.6"
django-filter = "^2.4.0"
python-dotenv = "^0.17.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
